Apply Eric's next-beta.patch that enables 64 bit FileIds and  adds new columns to the catalog.
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@8478 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/bc_types.h
@@ -57,7 +57,7 @@
  * In principle, the only field that really should need to be
  *  64 bits is the FileId_t
  */
-typedef uint32_t FileId_t;
+typedef uint64_t FileId_t;
 typedef uint32_t DBId_t;              /* general DB id type */
 typedef uint32_t JobId_t;
